+++ js/render/painter.js
@@ -10,15 +10,14 @@ var FrameHistory = require('./frame_history');
  *
  * @param {Canvas} gl an experimental-webgl drawing context
  */
-module.exports = GLPainter;
-function GLPainter(gl, transform) {
+module.exports = Painter;
+function Painter(gl, transform) {
     this.gl = glutil.extend(gl);
     this.transform = transform;
 
     this.reusableTextures = {};
     this.preFbos = {};
 
-    this.tileExtent = 4096;
     this.frameHistory = new FrameHistory();
 
     this.setup();
@@ -28,7 +27,7 @@ function GLPainter(gl, transform) {
  * Update the GL viewport, projection matrix, and transforms to compensate
  * for a new width and height value.
  */
-GLPainter.prototype.resize = function(width, height) {
+Painter.prototype.resize = function(width, height) {
     var gl = this.gl;
 
     this.width = width * browser.devicePixelRatio;
@@ -38,7 +37,7 @@ GLPainter.prototype.resize = function(width, height) {
 };
 
 
-GLPainter.prototype.setup = function() {
+Painter.prototype.setup = function() {
     var gl = this.gl;
 
     gl.verbose = true;
@@ -116,35 +115,60 @@ GLPainter.prototype.setup = function() {
     gl.bindBuffer(gl.ARRAY_BUFFER, this.backgroundBuffer);
     gl.bufferData(gl.ARRAY_BUFFER, new Int16Array([-1, -1, 1, -1, -1, 1, 1, 1]), gl.STATIC_DRAW);
 
-    // The tileExtentBuffer is used when drawing to a full *tile*
-    this.tileExtentBuffer = gl.createBuffer();
-    this.tileExtentBuffer.itemSize = 4;
-    this.tileExtentBuffer.itemCount = 4;
-    gl.bindBuffer(gl.ARRAY_BUFFER, this.tileExtentBuffer);
-    gl.bufferData(gl.ARRAY_BUFFER, new Int16Array([
-        // tile coord x, tile coord y, texture coord x, texture coord y
-                      0, 0,                    0, 0,
-        this.tileExtent, 0,                32767, 0,
-                      0, this.tileExtent,      0, 32767,
-        this.tileExtent, this.tileExtent,  32767, 32767
-    ]), gl.STATIC_DRAW);
-
-    // The debugBuffer is used to draw tile outlines for debugging
-    this.debugBuffer = gl.createBuffer();
-    this.debugBuffer.itemSize = 2;
-    this.debugBuffer.itemCount = 5;
-    gl.bindBuffer(gl.ARRAY_BUFFER, this.debugBuffer);
-    gl.bufferData(gl.ARRAY_BUFFER, new Int16Array([0, 0, 4095, 0, 4095, 4095, 0, 4095, 0, 0]), gl.STATIC_DRAW);
+    this.setExtent(4096);
 
     // The debugTextBuffer is used to draw tile IDs for debugging
     this.debugTextBuffer = gl.createBuffer();
     this.debugTextBuffer.itemSize = 2;
 };
 
+/**
+ * Rebind the necessary buffers to render at a different extent than
+ * the current one. No-ops if the extent is not changing.
+ *
+ * @param {number} newExtent
+ * @example
+ * this.setExtent(4096);
+ */
+Painter.prototype.setExtent = function(newExtent) {
+    if (!newExtent || newExtent === this.tileExtent) return;
+
+    this.tileExtent = newExtent;
+
+    var gl = this.gl;
+
+    // The tileExtentBuffer is used when drawing to a full *tile*
+    this.tileExtentBuffer = gl.createBuffer();
+    this.tileExtentBuffer.itemSize = 4;
+    this.tileExtentBuffer.itemCount = 4;
+    gl.bindBuffer(gl.ARRAY_BUFFER, this.tileExtentBuffer);
+    gl.bufferData(
+        gl.ARRAY_BUFFER,
+        new Int16Array([
+            // tile coord x, tile coord y, texture coord x, texture coord y
+            0, 0, 0, 0,
+            this.tileExtent, 0, 32767, 0,
+            0, this.tileExtent, 0, 32767,
+            this.tileExtent, this.tileExtent,  32767, 32767
+        ]),
+        gl.STATIC_DRAW);
+
+    // The debugBuffer is used to draw tile outlines for debugging
+    this.debugBuffer = gl.createBuffer();
+    this.debugBuffer.itemSize = 2;
+    this.debugBuffer.itemCount = 5;
+    gl.bindBuffer(gl.ARRAY_BUFFER, this.debugBuffer);
+    gl.bufferData(
+        gl.ARRAY_BUFFER,
+        new Int16Array([
+            0, 0, this.tileExtent - 1, 0, this.tileExtent - 1, this.tileExtent - 1, 0, this.tileExtent - 1, 0, 0]),
+        gl.STATIC_DRAW);
+};
+
 /*
  * Reset the color buffers of the drawing canvas.
  */
-GLPainter.prototype.clearColor = function() {
+Painter.prototype.clearColor = function() {
     var gl = this.gl;
     gl.clearColor(0, 0, 0, 0);
     gl.clear(gl.COLOR_BUFFER_BIT);
@@ -154,14 +178,14 @@ GLPainter.prototype.clearColor = function() {
  * Reset the drawing canvas by clearing the stencil buffer so that we can draw
  * new tiles at the same location, while retaining previously drawn pixels.
  */
-GLPainter.prototype.clearStencil = function() {
+Painter.prototype.clearStencil = function() {
     var gl = this.gl;
     gl.clearStencil(0x0);
     gl.stencilMask(0xFF);
     gl.clear(gl.STENCIL_BUFFER_BIT);
 };
 
-GLPainter.prototype.drawClippingMask = function(tile) {
+Painter.prototype.drawClippingMask = function(tile) {
     var gl = this.gl;
     gl.switchShader(this.fillShader, tile.posMatrix);
     gl.colorMask(false, false, false, false);
@@ -193,8 +217,8 @@ GLPainter.prototype.drawClippingMask = function(tile) {
 };
 
 // Overridden by headless tests.
-GLPainter.prototype.prepareBuffers = function() {};
-GLPainter.prototype.bindDefaultFramebuffer = function() {
+Painter.prototype.prepareBuffers = function() {};
+Painter.prototype.bindDefaultFramebuffer = function() {
     var gl = this.gl;
     gl.bindFramebuffer(gl.FRAMEBUFFER, null);
 };
@@ -210,7 +234,7 @@ var draw = {
     vertices: require('./draw_vertices')
 };
 
-GLPainter.prototype.render = function(style, options) {
+Painter.prototype.render = function(style, options) {
     this.style = style;
     this.options = options;
 
@@ -241,7 +265,8 @@ GLPainter.prototype.render = function(style, options) {
     }
 };
 
-GLPainter.prototype.drawTile = function(tile, layers) {
+Painter.prototype.drawTile = function(tile, layers) {
+    this.setExtent(tile.tileExtent);
     this.drawClippingMask(tile);
     this.drawLayers(layers, tile.posMatrix, tile);
 
@@ -250,7 +275,7 @@ GLPainter.prototype.drawTile = function(tile, layers) {
     }
 };
 
-GLPainter.prototype.drawLayers = function(layers, matrix, tile) {
+Painter.prototype.drawLayers = function(layers, matrix, tile) {
     for (var i = layers.length - 1; i >= 0; i--) {
         var layer = layers[i];
 
@@ -266,7 +291,7 @@ GLPainter.prototype.drawLayers = function(layers, matrix, tile) {
 };
 
 // Draws non-opaque areas. This is for debugging purposes.
-GLPainter.prototype.drawStencilBuffer = function() {
+Painter.prototype.drawStencilBuffer = function() {
     var gl = this.gl;
     gl.switchShader(this.fillShader, this.identityMatrix);
 
@@ -285,7 +310,7 @@ GLPainter.prototype.drawStencilBuffer = function() {
     gl.blendFunc(gl.ONE_MINUS_DST_ALPHA, gl.ONE);
 };
 
-GLPainter.prototype.translateMatrix = function(matrix, tile, translate, anchor) {
+Painter.prototype.translateMatrix = function(matrix, tile, translate, anchor) {
     if (!translate[0] && !translate[1]) return matrix;
 
     if (anchor === 'viewport') {
@@ -309,7 +334,7 @@ GLPainter.prototype.translateMatrix = function(matrix, tile, translate, anchor) 
     return translatedMatrix;
 };
 
-GLPainter.prototype.saveTexture = function(texture) {
+Painter.prototype.saveTexture = function(texture) {
     var textures = this.reusableTextures[texture.size];
     if (!textures) {
         this.reusableTextures[texture.size] = [texture];
@@ -319,7 +344,7 @@ GLPainter.prototype.saveTexture = function(texture) {
 };
 
 
-GLPainter.prototype.getTexture = function(size) {
+Painter.prototype.getTexture = function(size) {
     var textures = this.reusableTextures[size];
     return textures && textures.length > 0 ? textures.pop() : null;
 };