+++ js/symbol/shaping.js
@@ -137,13 +137,12 @@ function align(positionedGlyphs, justify, horizontalAlign, verticalAlign, maxLin
 function shapeIcon(image, layout) {
     if (!image || !image.rect) return null;
 
-    var padding = 1;
     var dx = layout['icon-offset'][0];
     var dy = layout['icon-offset'][1];
-    var x1 = dx - image.width / 2 - padding;
-    var x2 = x1 + image.rect.w;
-    var y1 = dy - image.height / 2 - padding;
-    var y2 = y1 + image.rect.h;
+    var x1 = dx - image.width / 2;
+    var x2 = x1 + image.width;
+    var y1 = dy - image.height / 2;
+    var y2 = y1 + image.height;
 
     return new PositionedIcon(image, y1, y2, x1, x2);
 }