+++ test/js/ui/map.test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var test = require('tape');
+var test = require('prova');
 var Map = require('../../../js/ui/map');
 var Style = require('../../../js/style/style');
 
@@ -91,81 +91,43 @@ test('Map', function(t) {
         });
     });
 
+    t.test('#resize', function(t) {
+        t.test('sets width and height from container offsets', function(t) {
+            var map = createMap(),
+                container = map.getContainer();
+
+            container.offsetWidth = 250;
+            container.offsetHeight = 250;
+            map.resize();
+
+            t.equal(map.transform.width, 250);
+            t.equal(map.transform.height, 250);
+
+            t.end();
+        });
+
+        t.test('fires movestart, move, resize, and moveend events', function(t) {
+            var map = createMap(),
+                events = [];
+
+            ['movestart', 'move', 'resize', 'moveend'].forEach(function (event) {
+                map.on(event, function(e) {
+                    events.push(e.type);
+                });
+            });
+
+            map.resize();
+            t.deepEqual(events, ['movestart', 'move', 'resize', 'moveend']);
+
+            t.end();
+        });
+    });
+
     t.test('#getBounds', function(t) {
         var map = createMap();
         t.deepEqual(parseFloat(map.getBounds().getCenter().lat.toFixed(10)), 0, 'getBounds');
         t.deepEqual(parseFloat(map.getBounds().getCenter().lng.toFixed(10)), 0, 'getBounds');
         t.end();
-    });
-
-    t.test('#setView', function(t) {
-        var map = createMap();
-
-        t.test('sets center', function(t) {
-            map.setView([1, 2], 3, 4);
-            t.deepEqual(map.getCenter(), { lat: 1, lng: 2 });
-            t.end();
-        });
-
-        t.test('sets zoom', function(t) {
-            map.setView([1, 2], 3, 4);
-            t.deepEqual(map.getZoom(), 3);
-            t.end();
-        });
-
-        t.test('sets bearing', function(t) {
-            map.setView([1, 2], 3, 4);
-            t.deepEqual(map.getBearing(), 4);
-            t.end();
-        });
-
-        t.test('emits move events', function(t) {
-            var started, ended;
-            map.on('movestart', function() { started = true; })
-                .on('moveend', function() { ended = true; });
-            map.setView([1, 2], 3, 4);
-            t.ok(started);
-            t.ok(ended);
-            t.end();
-        });
-
-        t.test('cancels in-progress easing', function(t) {
-            map.panTo([3, 4]);
-            t.ok(map.isEasing());
-            map.setView([1, 2], 3, 4);
-            t.ok(!map.isEasing());
-            t.end();
-        });
-
-        t.end();
-    });
-
-    t.test('#setCenter', function(t) {
-        var map = createMap();
-
-        t.test('sets center', function(t) {
-            map.setCenter([1, 2]);
-            t.deepEqual(map.getCenter(), { lat: 1, lng: 2 });
-            t.end();
-        });
-
-        t.test('emits move events', function(t) {
-            var started, ended;
-            map.on('movestart', function() { started = true; })
-                .on('moveend', function() { ended = true; });
-            map.setCenter([1, 2]);
-            t.ok(started);
-            t.ok(ended);
-            t.end();
-        });
-
-        t.test('cancels in-progress easing', function(t) {
-            map.panTo([3, 4]);
-            t.ok(map.isEasing());
-            map.setCenter([1, 2]);
-            t.ok(!map.isEasing());
-            t.end();
-        });
     });
 
     t.test('#remove', function(t) {
@@ -183,62 +145,6 @@ test('Map', function(t) {
             }
         };
         map.addControl(control);
-    });
-
-    t.test('#setZoom', function(t) {
-        var map = createMap();
-
-        t.test('sets zoom', function(t) {
-            map.setZoom(3);
-            t.deepEqual(map.getZoom(), 3);
-            t.end();
-        });
-
-        t.test('emits move events', function(t) {
-            var started, ended;
-            map.on('movestart', function() { started = true; })
-                .on('moveend', function() { ended = true; });
-            map.setZoom(3);
-            t.ok(started);
-            t.ok(ended);
-            t.end();
-        });
-
-        t.test('cancels in-progress easing', function(t) {
-            map.panTo([3, 4]);
-            t.ok(map.isEasing());
-            map.setZoom(3);
-            t.ok(!map.isEasing());
-            t.end();
-        });
-    });
-
-    t.test('#setBearing', function(t) {
-        var map = createMap();
-
-        t.test('sets bearing', function(t) {
-            map.setBearing(4);
-            t.deepEqual(map.getBearing(), 4);
-            t.end();
-        });
-
-        t.test('emits move events', function(t) {
-            var started, ended;
-            map.on('movestart', function() { started = true; })
-                .on('moveend', function() { ended = true; });
-            map.setBearing(4);
-            t.ok(started);
-            t.ok(ended);
-            t.end();
-        });
-
-        t.test('cancels in-progress easing', function(t) {
-            map.panTo([3, 4]);
-            t.ok(map.isEasing());
-            map.setBearing(4);
-            t.ok(!map.isEasing());
-            t.end();
-        });
     });
 
     t.test('#addClass', function(t) {