+++ CHANGELOG.md
@@ -28,13 +28,15 @@ An in-progress version being developed in the `master` branch.
 * Add map#setPitch(pitch) and map#getPitch()
 * Switch to JSDoc for documentation
 * Sort overlapping symbols by their y position (#1184)
+* Add `dblclick` event to Map. (#1168)
 
 #### Breaking
 
 * `mapboxgl.Source` is no longer exported. Use `map.addSource()` instead.
 * `mapboxgl.util.supported()` moved to `mapboxgl.supported()
-* map#setView(latlng, zoom, bearing) changed to map#setView(latlng, zoom, bearing, pitch)
-* map#easeTo(latlng, zoom, bearing, options) changed to map#easeTo(latlng, zoom, bearing, pitch, options)
+* map#setView(latlng, zoom, bearing) changed to map#jumpTo(options)
+* map#easeTo(latlng, zoom, bearing, options) changed to map#easeTo(options)
+* map#flyTo(latlng, zoom, bearing, options) changed to map#flyTo(options)